Move common classes for authorization to service base module.
