+++ server/src/main/java/org/eclipse/hono/authorization/impl/InMemoryAuthorizationService.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2016 Bosch Software Innovations GmbH.
+ * Copyright (c) 2016, 2017 Bosch Software Innovations GmbH.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -27,9 +27,9 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import java.util.stream.Collectors;
 
-import org.eclipse.hono.authorization.AccessControlList;
-import org.eclipse.hono.authorization.AclEntry;
-import org.eclipse.hono.authorization.Permission;
+import org.eclipse.hono.service.authorization.AccessControlList;
+import org.eclipse.hono.service.authorization.AclEntry;
+import org.eclipse.hono.service.authorization.Permission;
 import org.eclipse.hono.util.ResourceIdentifier;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
