+++ server/src/test/java/org/eclipse/hono/authorization/impl/InMemoryAuthorizationServiceTest.java
@@ -1,15 +1,26 @@
+/**
+ * Copyright (c) 2016, 2017 Bosch Software Innovations GmbH.
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *    Bosch Software Innovations GmbH - initial creation
+ */
 package org.eclipse.hono.authorization.impl;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import org.eclipse.hono.authorization.AuthorizationService;
-import org.eclipse.hono.authorization.Permission;
+import org.eclipse.hono.service.authorization.AuthorizationService;
+import org.eclipse.hono.service.authorization.Permission;
 import org.eclipse.hono.util.ResourceIdentifier;
 import org.junit.Before;
 import org.junit.Test;
 
 /**
- * Tests InMemoryAuthorizationServiceTest.
+ * Tests verifying bahvior of {@link InMemoryAuthorizationService}.
  */
 public class InMemoryAuthorizationServiceTest {
