+++ server/src/test/java/org/eclipse/hono/server/HonoServerTest.java
@@ -24,10 +24,10 @@ import org.apache.qpid.proton.amqp.transport.Target;
 import org.apache.qpid.proton.engine.Record;
 import org.apache.qpid.proton.engine.impl.RecordImpl;
 import org.eclipse.hono.TestSupport;
-import org.eclipse.hono.authorization.AuthorizationConstants;
-import org.eclipse.hono.authorization.Permission;
 import org.eclipse.hono.service.amqp.BaseEndpoint;
 import org.eclipse.hono.service.amqp.Endpoint;
+import org.eclipse.hono.service.authorization.AuthorizationConstants;
+import org.eclipse.hono.service.authorization.Permission;
 import org.eclipse.hono.telemetry.TelemetryConstants;
 import org.eclipse.hono.util.Constants;
 import org.eclipse.hono.util.ResourceIdentifier;
