+++ service-base/src/main/java/org/eclipse/hono/service/authorization/Permission.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2016 Bosch Software Innovations GmbH.
+ * Copyright (c) 2016, 2017 Bosch Software Innovations GmbH.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -9,12 +9,19 @@
  * Contributors:
  *    Bosch Software Innovations GmbH - initial API and implementation and initial documentation
  */
-package org.eclipse.hono.authorization;
+package org.eclipse.hono.service.authorization;
 
 /**
  * Defines permissions that are required to access Hono service resources.
  */
 public enum Permission
 {
-    READ, WRITE
+    /**
+     * Permission required for receiving message from a node.
+     */
+    READ,
+    /**
+     * Permission required for sending messages to a node.
+     */
+    WRITE
 }
